+++ frontend/src/app/components/screens/video-player/video-player.component.ts
@@ -14,9 +14,12 @@ export class VideoPlayerComponent {
     isActive = input.required<boolean>();
     header = input.required<string>();
     variables = input.required<any>();
+    loop = input.required<boolean>();
+    repeat = input.required<number>();
     finished = output<null>();
 
     media_id: string = '';
+    playCount: number = 0;
 
     constructor(
         private mediaService: MediaService
@@ -38,7 +41,10 @@ export class VideoPlayerComponent {
                 next: (media: Media) => {
                     var myVideo: any = document.getElementById("player");
                     this.media_id = media.id;
+                    this.playCount = 0;
                     myVideo.src = this.mediaService.getMediaUrl(media);
+                    if (this.loop()) myVideo.loop = true;
+                    else myVideo.loop = false;
                     myVideo.currentTime = 0;
                     this.startPlaying();
                 },
@@ -52,6 +58,17 @@ export class VideoPlayerComponent {
         if (this.isActive() && this.media_id != '') {
             var myVideo: any = document.getElementById("player");
             if (myVideo.paused) myVideo.play();
+        }
+    }
+
+    onVideoEnded() {
+        if (this.playCount < this.repeat()) {
+            this.playCount++;
+            var myVideo: any = document.getElementById("player");
+            myVideo.currentTime = 0;
+            myVideo.play();
+        } else {
+            this.finished.emit(null);
         }
     }
 }